writingToFile.java
So, this program attempts to write the specified content to a file named "output.txt" and informs you if the operation was successful or if any exceptions occurred during the process. Make sure you have the necessary file permissions and that the file "output.txt" does not exist before running this code, or it will be overwritten with the new content.